requirements.txt
torch
torchvision
numpy
matplotlib
tqdm
pyyaml
pillow
tensorboard
